// cgoml-ui/src/app/features/models/messaging/views/basic/OutboundMessageBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class OutboundMessageBasicView {

id: number;
selected: boolean;
canDelete: boolean;
uuid: string;
receptionDate: Date;
messageType: string;

/* Specific Code Start */
/* Specific Code End */
}