// cgoml-ui/src/app/features/models/messaging/filters/OutboundMessageFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class OutboundMessageFilter {

uuid: string;
receptionDateMinValue: Date;
receptionDateMaxValue: Date;
messageType: string;

/* Specific Code Start */
/* Specific Code End */
}