// cgoml-ui/src/app/features/models/messaging/sortings/OutboundMessageSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class OutboundMessageSorting {

uuidOrderType: string;
receptionDateOrderType: string;
messageTypeOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}